create metadata modal component, organize components
