+++ src/components/help/help-section.vue
