+++ src/components/panel-editors/chart-editor.vue
@@ -73,7 +73,7 @@
 </template>
 
 <script lang="ts">
-import ActionModal from '@/components/helpers/action-modal.vue';
+import ActionModal from '@/components/support/action-modal.vue';
 import { Options, Prop, Vue } from 'vue-property-decorator';
 import {
     ChartConfig,
@@ -84,8 +84,8 @@ import {
     SlideshowChartPanel,
     SourceCounts
 } from '@/definitions';
-import ChartPreviewV from './helpers/chart-preview.vue';
-import ConfirmationModalV from './helpers/confirmation-modal.vue';
+import ChartPreviewV from '../support/chart-preview.vue';
+import ConfirmationModalV from '../support/confirmation-modal.vue';
 import draggable from 'vuedraggable';
 
 @Options({
