+++ src/components/panel-editors/custom-editor.vue
