+++ src/components/panel-editors/slideshow-editor.vue
