+++ src/components/support/chart-preview.vue
