+++ src/components/support/loading-page.vue
