+++ src/components/support/time-slider-editor.vue
